Implemented --user-data-dir Argument
In this commit, the bot now stores the browser user data in a selenium folder within the current working directory. After a brief refactor the bot now detects if the user is already logged in to instagram and if so starts scraping. If the user isn't logged in, it logs them in and then starts scraping. To prevent unnecessary bloat in the repo I've added the selenium directory to the gitignore file.